algorithm_globals.h, renamed it and split it by moving the version numbers out.
