+++ includes/acl/algorithm/uniformly_sampled/encoder.h
@@ -27,8 +27,8 @@
 #include "acl/core/memory.h"
 #include "acl/core/error.h"
 #include "acl/core/bitset.h"
-#include "acl/core/algorithm_globals.h"
 #include "acl/core/utils.h"
+#include "acl/core/algorithm_types.h"
 #include "acl/algorithm/uniformly_sampled/common.h"
 #include "acl/compression/compressed_clip_impl.h"
 #include "acl/compression/skeleton.h"
