+++ includes/acl/core/ialgorithm.h
@@ -25,8 +25,8 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include "acl/core/memory.h"
-#include "acl/core/algorithm_globals.h"
 #include "acl/core/compressed_clip.h"
+#include "acl/core/algorithm_types.h"
 #include "acl/compression/skeleton.h"
 #include "acl/compression/animation_clip.h"
 #include "acl/decompression/output_writer.h"
